src/llm_analyzer/analyzer.py
"""
LLM Security Analyzer

Core component for semantic analysis of cloud-native security artifacts.
Achieves 94% accuracy through few-shot prompt engineering and schema validation.
"""

import json
import logging
from typing import Dict, List, Optional, Union
from dataclasses import dataclass
from enum import Enum

import openai
from openai import OpenAI
import yaml
from pydantic import BaseModel, Field

from .prompts import SecurityPromptTemplates
from .validators import ArtifactValidator


class ArtifactType(Enum):
    """Supported artifact types for security analysis."""
    DOCKERFILE = "dockerfile"
    KUBERNETES_YAML = "kubernetes_yaml"
    TERRAFORM = "terraform"
    CICD_PIPELINE = "cicd_pipeline"


class SecurityFinding(BaseModel):
    """Structured representation of a security finding."""
    finding_id: str = Field(..., description="Unique identifier for the finding")
    severity: str = Field(..., description="Severity level: CRITICAL, HIGH, MEDIUM, LOW")
    category: str = Field(..., description="Security category (e.g., 'Privilege Escalation', 'Secrets Management')")
    title: str = Field(..., description="Brief title of the security issue")
    description: str = Field(..., description="Detailed description of the vulnerability")
    line_number: Optional[int] = Field(None, description="Line number where issue occurs")
    remediation: str = Field(..., description="Specific remediation guidance")
    cwe_id: Optional[str] = Field(None, description="CWE identifier if applicable")
    mitre_technique: Optional[str] = Field(None, description="MITRE ATT&CK technique")
    confidence: float = Field(..., ge=0.0, le=1.0, description="Confidence score (0.0-1.0)")


class AnalysisResult(BaseModel):
    """Complete analysis result for an artifact."""
    artifact_type: ArtifactType
    artifact_content: str
    findings: List[SecurityFinding]
    overall_risk_score: float = Field(..., ge=0.0, le=10.0)
    analysis_metadata: Dict = Field(default_factory=dict)


class LLMSecurityAnalyzer:
    """
    LLM-powered security analyzer for cloud-native artifacts.
    
    Features:
    - Few-shot prompt engineering for high accuracy
    - Schema validation for structured output
    - Support for multiple artifact types
    - Confidence scoring and fallback mechanisms
    """
    
    def __init__(
        self,
        api_key: str,
        model: str = "gpt-4",
        temperature: float = 0.1,
        max_tokens: int = 2048
    ):
        self.client = OpenAI(api_key=api_key)
        self.model = model
        self.temperature = temperature
        self.max_tokens = max_tokens
        self.prompt_templates = SecurityPromptTemplates()
        self.validator = ArtifactValidator()
        self.logger = logging.getLogger(__name__)
        
    def analyze_artifact(
        self,
        content: str,
        artifact_type: ArtifactType,
        context: Optional[Dict] = None
    ) -> AnalysisResult:
        """
        Analyze a security artifact for vulnerabilities and misconfigurations.
        
        Args:
            content: Raw content of the artifact
            artifact_type: Type of artifact being analyzed
            context: Additional context for analysis
            
        Returns:
            AnalysisResult with structured findings
        """
        try:
            # Validate artifact format
            validation_result = self.validator.validate_artifact(content, artifact_type)
            if not validation_result.is_valid:
                raise ValueError(f"Invalid artifact format: {validation_result.errors}")
            
            # Get appropriate prompt template
            prompt = self.prompt_templates.get_analysis_prompt(
                artifact_type=artifact_type,
                content=content,
                context=context or {}
            )
            
            # Perform LLM analysis
            response = self.client.chat.completions.create(
                model=self.model,
                messages=[
                    {"role": "system", "content": prompt.system_message},
                    {"role": "user", "content": prompt.user_message}
                ],
                temperature=self.temperature,
                max_tokens=self.max_tokens,
                response_format={"type": "json_object"}
            )
            
            # Parse structured response
            analysis_data = json.loads(response.choices[0].message.content)
            
            # Convert to structured findings
            findings = [
                SecurityFinding(**finding) for finding in analysis_data.get("findings", [])
            ]
            
            # Calculate overall risk score
            risk_score = self._calculate_risk_score(findings)
            
            return AnalysisResult(
                artifact_type=artifact_type,
                artifact_content=content,
                findings=findings,
                overall_risk_score=risk_score,
                analysis_metadata={
                    "model_used": self.model,
                    "analysis_timestamp": validation_result.timestamp,
                    "token_usage": response.usage.dict() if response.usage else {},
                    "validation_warnings": validation_result.warnings
                }
            )
            
        except Exception as e:
            self.logger.error(f"Analysis failed for {artifact_type.value}: {str(e)}")
            raise
    
    def batch_analyze(
        self,
        artifacts: List[Dict[str, Union[str, ArtifactType]]],
        parallel: bool = True
    ) -> List[AnalysisResult]:
        """
        Analyze multiple artifacts in batch.
        
        Args:
            artifacts: List of artifacts with content and type
            parallel: Whether to process in parallel
            
        Returns:
            List of analysis results
        """
        results = []
        
        for artifact in artifacts:
            try:
                result = self.analyze_artifact(
                    content=artifact["content"],
                    artifact_type=artifact["type"],
                    context=artifact.get("context")
                )
                results.append(result)
            except Exception as e:
                self.logger.error(f"Batch analysis failed for artifact: {str(e)}")
                # Add empty result to maintain order
                results.append(None)
                
        return results
    
    def _calculate_risk_score(self, findings: List[SecurityFinding]) -> float:
        """Calculate overall risk score based on findings."""
        if not findings:
            return 0.0
            
        severity_weights = {
            "CRITICAL": 4.0,
            "HIGH": 3.0,
            "MEDIUM": 2.0,
            "LOW": 1.0
        }
        
        weighted_score = sum(
            severity_weights.get(finding.severity, 1.0) * finding.confidence
            for finding in findings
        )
        
        # Normalize to 0-10 scale
        max_possible_score = len(findings) * 4.0  # All critical with 100% confidence
        normalized_score = min(10.0, (weighted_score / max_possible_score) * 10.0)
        
        return round(normalized_score, 2)
    
    def get_educational_hints(
        self,
        finding: SecurityFinding,
        student_level: str = "intermediate"
    ) -> List[str]:
        """
        Generate educational hints for a specific finding.
        
        Args:
            finding: Security finding to generate hints for
            student_level: Skill level of the student
            
        Returns:
            List of progressive hints
        """
        hint_prompt = self.prompt_templates.get_hint_prompt(finding, student_level)
        
        response = self.client.chat.completions.create(
            model=self.model,
            messages=[
                {"role": "system", "content": hint_prompt.system_message},
                {"role": "user", "content": hint_prompt.user_message}
            ],
            temperature=0.2,
            max_tokens=1024
        )
        
        hints_data = json.loads(response.choices[0].message.content)
        return hints_data.get("hints", [])
