requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning & AI
openai==1.3.7
anthropic==0.7.8
transformers==4.36.0
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Reinforcement Learning
stable-baselines3==2.2.1
gymnasium==0.29.1
tensorboard==2.15.1
matplotlib==3.8.2

# Kubernetes & Container Management
kubernetes==28.1.0
docker==6.1.3
pyyaml==6.0.1
helm==0.1.0

# Security Analysis
bandit==1.7.5
safety==2.3.5
semgrep==1.45.0
trivy-python==0.1.0

# Database & Storage
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Monitoring & Observability
prometheus-client==0.19.0
jaeger-client==4.8.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
testcontainers==3.7.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
